[UNFINISHED] First iteration of improved scrolling.
I began to modularize the components of scrolling into: EventBuffer, PanRecognizer, and LayerDraggable2 (...this is clearly very work-in-progress). Perhaps there is a less aggressive approach that we could get into the product faster and with fewer changes... though we should keep in mind the complexity around nested draggables and tap/drag event conflicts as these are hard to add cleanly later.