Update resource optimization scripts to use ImageMagick and OpenJPEG to handle JP2 compression.
This fixes compatibility issues with JP2 files on macOS 10.12.2 (and issues with JP2 files created on 10.12.2 with all previous macOS versions).

Note that we need take a step via BMP to generate the JP2 because OpenJPEG won’t deal with the color profile embedded in the PNGs. ImageMagick can convert PNGs to BMPs with an alpha channel, while the sips tool cannot.