Refactor ancestors() into containers()
ancestors() used to be a weird API that had an argument ‘context’ that would let you bubble up to the root. I can’t imagine a scenario where you would want this in a prototype, so that argument has been removed, and for internal use, the API has been replaced by `containers()`.